# Conflicts:
#	resources/vanilla
#	src/pocketmine/Player.php

+++ src/network/mcpe/handler/InGamePacketHandler.php
@@ -659,6 +659,9 @@ class InGamePacketHandler extends PacketHandler{
 				$modifiedPages[] = $packet->pageNumber;
 				break;
 			case BookEditPacket::TYPE_DELETE_PAGE:
+				if(!$newBook->pageExists($packet->pageNumber)){
+					return false;
+				}
 				$newBook->deletePage($packet->pageNumber);
 				$modifiedPages[] = $packet->pageNumber;
 				break;
